refactor(editor): move embed icons to icons.tsx and revert env package modifications
