+++ .gitignore
@@ -10,8 +10,6 @@ build
 
 # Generated files
 apps/web/src/routeTree.gen.ts
-AGENTS.md
-CLAUDE.md
 
 # Environment variables
 .env
