+++ packages/editor/src/extensions/resizable-node-view.tsx
@@ -3,6 +3,12 @@
 import { NodeViewWrapper } from "@tiptap/react";
 import type { NodeViewProps } from "@tiptap/react";
 import { useCallback, useEffect, useRef, useState } from "react";
+import {
+  AlignIcon,
+  DeleteEmbedIcon,
+  EditUrlIcon,
+  OpenInYouTubeIcon,
+} from "../icons";
 
 interface ResizableNodeViewProps extends NodeViewProps {
   children: React.ReactNode;
@@ -16,47 +22,6 @@ interface ResizableNodeViewProps extends NodeViewProps {
 }
 
 type ResizeDirection = "left" | "right" | "bottom" | "corner-bl" | "corner-br";
-
-type AlignDir = "left" | "center" | "right";
-
-const ALIGN_PATHS: Record<AlignDir, React.ReactNode> = {
-  center: (
-    <>
-      <line x1="21" y1="6" x2="3" y2="6" />
-      <line x1="17" y1="12" x2="7" y2="12" />
-      <line x1="19" y1="18" x2="5" y2="18" />
-    </>
-  ),
-  left: (
-    <>
-      <line x1="21" y1="6" x2="3" y2="6" />
-      <line x1="15" y1="12" x2="3" y2="12" />
-      <line x1="17" y1="18" x2="3" y2="18" />
-    </>
-  ),
-  right: (
-    <>
-      <line x1="21" y1="6" x2="3" y2="6" />
-      <line x1="21" y1="12" x2="9" y2="12" />
-      <line x1="21" y1="18" x2="7" y2="18" />
-    </>
-  ),
-};
-
-const AlignIcon = ({ dir }: { dir: AlignDir }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 24 24"
-    fill="none"
-    stroke="currentColor"
-    strokeWidth="2"
-    strokeLinecap="round"
-    strokeLinejoin="round"
-    className="rte-editor-icon"
-  >
-    {ALIGN_PATHS[dir]}
-  </svg>
-);
 
 interface EmbedToolbarProps {
   align: string;
@@ -172,19 +137,7 @@ const EmbedToolbar = ({
               setIsEditingUrl(true);
             }}
           >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="rte-editor-icon"
-            >
-              <path d="M12 20h9" />
-              <path d="M16.5 3.5a2.121 2.121 0 0 1 3 3L7 19l-4 1 1-4L16.5 3.5z" />
-            </svg>
+            <EditUrlIcon />
           </button>
         )}
         {watchUrl && (
@@ -195,20 +148,7 @@ const EmbedToolbar = ({
             className="rte-embed-selector-btn"
             onClick={() => window.open(watchUrl, "_blank")}
           >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="rte-editor-icon"
-            >
-              <path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6" />
-              <polyline points="15 3 21 3 21 9" />
-              <line x1="10" y1="14" x2="21" y2="3" />
-            </svg>
+            <OpenInYouTubeIcon />
           </button>
         )}
         {deleteNode && (
@@ -219,19 +159,7 @@ const EmbedToolbar = ({
             className="rte-embed-selector-btn rte-embed-selector-btn--danger"
             onClick={() => deleteNode()}
           >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 24 24"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              className="rte-editor-icon"
-            >
-              <polyline points="3 6 5 6 21 6" />
-              <path d="M19 6v14a2 2 0 0 1-2 2H7a2 2 0 0 1-2-2V6m3 0V4a2 2 0 0 1 2-2h4a2 2 0 0 1 2 2v2" />
-            </svg>
+            <DeleteEmbedIcon />
           </button>
         )}
       </div>
